# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task API",
        "description": "API para gestionar tareas.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Obtiene una lista de tareas.",
                "description": "Lista de tareas.",
                "operationId": "471d938a3bfaeb6fbd981d7c3e2728b0",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado (all, P, E, C).",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de tareas obtenida exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado."
                    }
                }
            },
            "post": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Almacena una nueva tarea.",
                "description": "Almacenar una nueva tarea en el almacenamiento.",
                "operationId": "f803ebdb2300acdb7e93bbe8d0b9a2e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarea creada exitosamente."
                    },
                    "422": {
                        "description": "Datos de entrada no válidos."
                    }
                }
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Obtiene los detalles de una tarea específica.",
                "description": "Mostrar detalles de una tarea específica.",
                "operationId": "90b3cd4d2d4b416dc989d2055dbed06c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarea.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la tarea obtenidos exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tarea no encontrada."
                    }
                }
            },
            "put": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Actualiza una tarea específica.",
                "description": "Actualizar una tarea específica en el almacenamiento.",
                "operationId": "1aaf37e74360ef4c10fadaee02ae77fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarea a actualizar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ModifyTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarea actualizada exitosamente."
                    },
                    "404": {
                        "description": "Tarea no encontrada."
                    },
                    "422": {
                        "description": "Datos de entrada no válidos."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Elimina una tarea específica.",
                "description": "Eliminar una tarea específica.",
                "operationId": "ad582480cd57c2a4e3b6a426943eeca5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarea a eliminar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarea eliminada exitosamente."
                    },
                    "404": {
                        "description": "Tarea no encontrada."
                    }
                }
            }
        },
        "/api/tasks/{id}/complete": {
            "get": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Muestra el formulario para completar una tarea.",
                "description": "Mostrar formulario para completar una tarea.",
                "operationId": "bf22649c8295acbdf5b85d24c2b544b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarea a completar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Formulario para completar la tarea mostrado exitosamente."
                    },
                    "404": {
                        "description": "Tarea no encontrada."
                    }
                }
            },
            "post": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Completa una tarea específica.",
                "description": "Completar una tarea.",
                "operationId": "93544830d9b827807253b65ef8b7d88d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarea a completar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompleteTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarea completada exitosamente."
                    },
                    "404": {
                        "description": "Tarea no encontrada."
                    }
                }
            }
        },
...
        "/api/tasks/{id}/edit": {
            "get": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Muestra el formulario para editar una tarea existente.",
                "description": "Mostrar formulario para editar una tarea existente.",
                "operationId": "0c6e9b9518c3380edab73353c6dd5849",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarea a editar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Formulario para editar la tarea mostrado exitosamente."
                    },
                    "404": {
                        "description": "Tarea no encontrada."
                    }
                }
            }
        },
        "/api/tasks/client": {
            "post": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Almacena una nueva tarea creada por un cliente.",
                "description": "Almacenar una nueva tarea para el cliente.",
                "operationId": "05019c9ac31be61bc4b134b352c73805",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewTaskClientRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarea creada por el cliente exitosamente."
                    },
                    "422": {
                        "description": "Datos de entrada no válidos."
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CompleteTaskRequest": {
                "title": "Complete Task Request",
                "description": "Request para completar una tarea existente.",
                "required": [
                    "status",
                    "subsequent_notes",
                    "summary_file"
                ],
                "properties": {
                    "status": {
                        "description": "Estado de la tarea (C: Completada, X: Cancelada).",
                        "type": "string"
                    },
                    "subsequent_notes": {
                        "description": "Notas posteriores a la realización de la tarea.",
                        "type": "string"
                    },
                    "summary_file": {
                        "description": "Archivo de resumen de la tarea (PDF, máximo 10MB).",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            },
...
            "ModifyTaskRequest": {
                "title": "Modify Task Request",
                "description": "Request para modificar una tarea existente.",
                "required": [
                    "client",
                    "contact_person",
                    "contact_email",
                    "contact_phone",
                    "description",
                    "address",
                    "city",
                    "postal_code",
                    "province",
                    "assigned_operator",
                    "status"
                ],
                "properties": {
                    "client": {
                        "description": "Nombre del cliente.",
                        "type": "string"
                    },
                    "contact_person": {
                        "description": "Persona de contacto.",
                        "type": "string"
                    },
                    "contact_email": {
                        "description": "Correo electrónico de contacto.",
                        "type": "string",
                        "format": "email"
                    },
                    "contact_phone": {
                        "description": "Teléfono de contacto (9 dígitos).",
                        "type": "string"
                    },
                    "description": {
                        "description": "Descripción de la tarea.",
                        "type": "string"
                    },
                    "address": {
                        "description": "Dirección.",
                        "type": "string"
                    },
                    "city": {
                        "description": "Ciudad.",
                        "type": "string"
                    },
                    "postal_code": {
                        "description": "Código postal (5 dígitos).",
                        "type": "string"
                    },
                    "province": {
                        "description": "Provincia.",
                        "type": "string"
                    },
                    "assigned_operator": {
                        "description": "ID del operador asignado.",
                        "type": "integer"
                    },
                    "status": {
                        "description": "Estado de la tarea (P, E, R, C, X).",
                        "type": "string"
                    },
                    "realization_date": {
                        "description": "Fecha de realización (opcional).",
                        "type": "string",
                        "format": "date"
                    },
                    "previous_notes": {
                        "description": "Notas previas (opcional).",
                        "type": "string"
                    },
                    "subsequent_notes": {
                        "description": "Notas posteriores (opcional).",
                        "type": "string"
                    },
                    "summary_file": {
                        "description": "Archivo de resumen (PDF, máximo 10MB, opcional).",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            },
...
            "StoreTaskRequest": {
                "title": "Store Task Request",
                "description": "Request para almacenar una nueva tarea.",
                "required": [
                    "client",
                    "contact_person",
                    "description",
                    "contact_phone",
                    "province",
                    "contact_email",
                    "status",
                    "address",
                    "city",
                    "assigned_operator",
                    "previous_notes"
                ],
                "properties": {
                    "client": {
                        "description": "Nombre del cliente.",
                        "type": "string"
                    },
                    "contact_person": {
                        "description": "Persona de contacto.",
                        "type": "string"
                    },
                    "description": {
                        "description": "Descripción de la tarea.",
                        "type": "string"
                    },
                    "contact_phone": {
                        "description": "Teléfono de contacto (9 dígitos).",
                        "type": "string"
                    },
                    "postal_code": {
                        "description": "Código postal (5 dígitos, opcional).",
                        "type": "string"
                    },
                    "province": {
                        "description": "Código de la provincia.",
                        "type": "string"
                    },
                    "contact_email": {
                        "description": "Correo electrónico de contacto.",
                        "type": "string",
                        "format": "email"
                    },
                    "status": {
                        "description": "Estado de la tarea.",
                        "type": "string"
                    },
                    "address": {
                        "description": "Dirección.",
                        "type": "string"
                    },
                    "city": {
                        "description": "Ciudad.",
                        "type": "string"
                    },
                    "assigned_operator": {
                        "description": "ID del operador asignado.",
                        "type": "integer"
                    },
                    "previous_notes": {
                        "description": "Notas previas (opcional).",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Task": {
                "title": "Task",
                "description": "Modelo de tarea.",
                "properties": {
                    "id": {
                        "description": "ID de la tarea.",
                        "type": "integer"
                    },
                    "client": {
                        "description": "Nombre del cliente.",
                        "type": "string"
                    },
                    "contact_person": {
                        "description": "Persona de contacto.",
                        "type": "string"
                    },
                    "contact_email": {
                        "description": "Email de contacto.",
                        "type": "string"
                    },
                    "contact_phone": {
                        "description": "Teléfono de contacto.",
                        "type": "string"
                    },
                    "description": {
                        "description": "Descripción de la tarea.",
                        "type": "string"
                    },
                    "address": {
                        "description": "Dirección.",
                        "type": "string"
                    },
                    "city": {
                        "description": "Ciudad.",
                        "type": "string"
                    },
                    "postal_code": {
                        "description": "Código postal.",
                        "type": "string"
                    },
                    "province": {
                        "description": "ID de la provincia.",
                        "type": "integer"
                    },
                    "assigned_operator": {
                        "description": "ID del operador asignado.",
                        "type": "integer"
                    },
                    "status": {
                        "description": "Estado de la tarea.",
                        "type": "string"
                    },
                    "realization_date": {
                        "description": "Fecha de realización.",
                        "type": "string",
                        "format": "date-time"
                    },
                    "previous_notes": {
                        "description": "Notas previas.",
                        "type": "string"
                    },
                    "subsequent_notes": {
                        "description": "Notas posteriores.",
                        "type": "string"
                    },
                    "summary_file": {
                        "description": "Ruta del archivo de resumen.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}